Versao 1 para o teste em production

+++ src/infrastructure/database/models/IndividualEvaluation.ts
@@ -61,7 +61,7 @@ IndividualEvaluation.init({
     allowNull: true,
   },
   totalScore: {
-    type: DataTypes.INTEGER,
+    type: DataTypes.DECIMAL(10, 1),
     allowNull: false,
     defaultValue: 0,
   },
