+++ src/infrastructure/database/models/IndividualRanking.ts
@@ -44,7 +44,7 @@ IndividualRanking.init({
     allowNull: false,
   },
   totalScore: {
-    type: DataTypes.INTEGER,
+    type: DataTypes.DECIMAL(10, 1),
     allowNull: false,
     defaultValue: 0,
   },
